# shared/openapi/mesob-api.yaml
openapi: "3.1.0"
info:
  title: Mesob Wallet API
  version: "1.0.0"
  description: |
    Unified public API for the Mesob Wallet platform.
    All requests go through the API Gateway at port 8000.

    ## Authentication
    - Mobile/USSD users authenticate with **MSISDN + PIN** via `/v1/identity/login`
    - Access tokens are RS256 JWTs valid for **15 minutes**
    - Refresh tokens are 7-day opaque tokens returned as an `httpOnly` cookie
    - Staff users (ADMIN, BRANCH_MANAGER, SUPER_ADMIN) use the same auth flow

    ## Demo Accounts
    | Role | MSISDN | PIN |
    |------|--------|-----|
    | SUPER_ADMIN | +251911000001 | 111111 |
    | ADMIN | +251911000002 | 111111 |
    | BRANCH_MANAGER | +251911000003 | 111111 |
    | AGENT | +251911000004 | 111111 |
    | USER (KYC Tier 2) | +251911000005 | 111111 |
    | USER (KYC Tier 0) | +251911000006 | 111111 |

    ## Amounts
    All monetary values are in **ETB minor units (cents)**.
    Example: 100.00 ETB = `10000`

    ## Response Envelope
    All responses follow:
    ```json
    {
      "data": { ... },
      "meta": { "request_id": "...", "trace_id": "..." }
    }
    ```
    Errors:
    ```json
    {
      "error": { "code": "DOMAIN_ERROR_CODE", "message": "human message" },
      "meta": { ... }
    }
    ```
  contact:
    name: Mesob Engineering
    email: engineering@mesobwallet.et
servers:
  - url: http://localhost:8000/v1
    description: Local development
  - url: https://api.mesobwallet.et/v1
...
paths:
...
  /agent/cash-in:
    post:
      tags: [Agent]
      summary: Agent disburses cash to user wallet
      operationId: agentCashIn
      security:
        - bearerAuth: []
      description: Requires `AGENT` role. Debits agent float, credits user wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_msisdn, amount_minor, idempotency_key]
              properties:
                user_msisdn:
                  type: string
                  example: "+251912345678"
                amount_minor:
                  type: integer
                  format: int64
                  example: 50000
                idempotency_key:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Cash-in posted
        "402":
          description: Insufficient float
        "403":
          $ref: "#/components/responses/Forbidden"
  /agent/cash-out:
    post:
      tags: [Agent]
      summary: Agent collects cash from user wallet
      operationId: agentCashOut
      security:
        - bearerAuth: []
      description: Requires `AGENT` role. Debits user wallet, credits agent float.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_msisdn, amount_minor, user_pin_confirmation, idempotency_key]
              properties:
                user_msisdn:
                  type: string
                  example: "+251912345678"
                amount_minor:
                  type: integer
                  format: int64
                  example: 30000
                user_pin_confirmation:
                  type: string
                  description: User's PIN to authorize the debit
                  example: "111111"
                idempotency_key:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Cash-out posted
        "402":
          description: Insufficient user balance
        "403":
          $ref: "#/components/responses/Forbidden"
  /agent/sync:
    post:
      tags: [Agent]
      summary: Batch sync offline operations
      operationId: agentSync
      security:
        - bearerAuth: []
      description: Processes a batch of offline operations captured with UUIDv7 keys.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operations]
              properties:
                operations:
                  type: array
                  maxItems: 100
                  items:
                    type: object
                    required: [idempotency_key, type, payload, captured_at]
                    properties:
                      idempotency_key:
                        type: string
                        format: uuid
                      type:
                        type: string
                        enum: [CASH_IN, CASH_OUT, ONBOARD]
                      payload:
                        type: object
                      captured_at:
                        type: string
                        format: date-time
      responses:
        "200":
          description: Sync result with per-operation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
              example:
                data:
                  accepted: 3
                  rejected: 0
                  results:
                    - idempotency_key: "01924b6f-1111-7000-8000-000000000001"
                      status: "ACCEPTED"
  /agent/onboard:
    post:
      tags: [Agent]
      summary: Agent onboards a new user (offline-capable)
      operationId: agentOnboard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [msisdn, full_name, idempotency_key]
              properties:
                msisdn:
                  type: string
                  example: "+251912999888"
                full_name:
                  type: string
                  example: "Tigist Haile"
                region_id:
                  type: string
                  format: uuid
                idempotency_key:
                  type: string
                  format: uuid
      responses:
        "201":
          description: User onboarded

  # ── BRANCH ────────────────────────────────────────────────────────────────
...
  /fraud/alerts:
    get:
      tags: [Fraud]
      summary: List open fraud alerts
      operationId: listFraudAlerts
      security:
        - bearerAuth: []
      description: Requires `ADMIN` or `SUPER_ADMIN` role.
      parameters:
        - in: query
          name: severity
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH]
        - in: query
          name: status
          schema:
            type: string
            enum: [OPEN, DISMISSED, ESCALATED]
            default: OPEN
        - $ref: "#/components/parameters/QueryLimit"
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
              example:
                data:
                  - alert_id: "alert_01J3ABCDEF"
                    severity: "HIGH"
                    rule: "VELOCITY_BREACH"
                    transaction_id: "01924b6f-1111-7000-8000-000000000001"
                    status: "OPEN"
                    created_at: "2026-05-31T09:00:00Z"
  /fraud/alerts/{id}/disposition:
    patch:
      tags: [Fraud]
      summary: Dismiss or escalate a fraud alert
      operationId: disposeFraudAlert
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [DISMISS, ESCALATE]
                notes:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Alert disposition saved

  # ── USSD ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT access token. Obtain via /identity/login or /identity/token/refresh.
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
...
    QueryLimit:
      in: query
      name: limit
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        data:
          description: Response payload (schema varies by endpoint)
        meta:
          $ref: "#/components/schemas/MetaBlock"
      required: [data]
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "INSUFFICIENT_BALANCE"
            message:
              type: string
              description: Human-readable message
              example: "Your balance is too low for this transaction"
        meta:
          $ref: "#/components/schemas/MetaBlock"
      required: [error]
    MetaBlock:
      type: object
      properties:
        request_id:
          type: string
          example: "req_01J3ABCDEF"
        trace_id:
          type: string
          example: "trace_01J3ABCDEF"
...
  responses:
...
    Forbidden:
      description: Insufficient role/permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: "FORBIDDEN"
              message: "This action requires BRANCH_MANAGER role"
...